+++ releasenotes.md
@@ -8,7 +8,7 @@
 * The branding in the console has been redesigned.
 * The following changes have been made to the support for John Romero’s megawad [*SIGIL*](http://SI6IL.com/):
   * `SIGIL.wad` will now only be automatically loaded alongside *The Ultimate DOOM* and not the registered version of *DOOM*.
-  * If [*Buckethead’s*](http://www.bucketheadpikes.com/) `SIGIL_SHREDS.wad` is found in the same folder as `SIGIL.wad`, it will now also be automatically loaded.
+  * If [*Buckethead’s*](https://en.wikipedia.org/wiki/Buckethead) `SIGIL_SHREDS.wad` is found, it will now also be automatically loaded.-
   * Savegame descriptions will now be updated with the current map’s name when saving over an existing savegame.
   * The `IDMUS` cheat can now be used to play *SIGIL’s* music.
   * The `episode` CVAR will now be updated if the `IDCLEV` cheat is used to warp to a map in *SIGIL*.
@@ -46,6 +46,7 @@
   * Also fade to black when quitting *DOOM Retro*.
   * No longer be negatively affected by the `vid_vsync` CVAR.
   * Be displayed correctly when the help screen is opened on the main screen by pressing the <kbd>F1</kbd> key.
+  * Still be created if there’s only one display. This will allow it to appear in a separate window when the `vid_fullscreen` CVAR is `off`.
 * Entering the `IDBEHOLD` cheat without any parameter will now:
   * Timeout after 2 seconds like all the other cheats.
   * Show underscores under the message if entered when the alternate widescreen HUD is displayed.